Python script to run sol-fbp-generator
This script is intended to run at compile time and will call
sol-fbp-generator with suitable parameters to generate code
for modules (like calamari) that depend on subflows. It will
search for .fbp files within each module directory and call
sol-fbp-generator to create a C file per fbp found with the
necessary code to get a node type.

Names for child_opts_set callback and the function that returns
the node type itself will be based on fbp file names. This is so
because we need a means to differentiate this assets in files with
more than one subflow, which is the case for calamari.c that
describes calamari/7seg and calamari/rgb-led.

Signed-off-by: Luiz Ywata <luizg.ywata@intel.com>