image-generator.sh
This script is based on arch-installer.sh. It generates a bootable
media with an Arch image. Now it also accepts the '-c' options, which
takes either an fbp or a C file as parameter, compiles and install it
into $ROOTFS/usr/bin and starts the program at boot time. All the
necessary libs to run a soletta program are also properly installed.

At this point this script must be run from soletta/tools directory
since all paths within it are relative to this location. Further
improvements may change this soon.

You can run the script using the simple.c file that is being
commited to the same directory. Unfortunately, sol-fbp-generator
is missing some #includes which I had to provide in order to make
the file compile. This is a problem that shall be addressed soon.